Separate V9 per-asset evaluation from aggregation
Extract single-asset scoring into evaluate-asset.ts, leaving cohort planning and common-mode aggregation in evaluate-set.ts with its export surface intact. Pinned scoring, golden, and candidate suites pass with zero fixture changes. Audit findings R-13/Q-06 (P3-04 tranche 3).
